Game 7 proj. question update

+++ app/workers/3D_models_worker.py
@@ -3,17 +3,27 @@
 from game7_worker import convert_stl_to_voxels
 import glob
 import numpy as np
+from os import path
+
+# Run this script to convert any new STLs to npy
 
 def convert_all_stls_to_npy():
-    print("Converting all stl files in static/data/solids folder to npy...")
+    print("Converting all stl files without npy counterparts in static/data/solids folder to npy...")
     print("This might take a while.")
     # Find all stls
     # Convert each and safe to npy
     all_stl_paths = glob.glob(f'./static/data/solids/*.stl')
     for ind in range(len(all_stl_paths)):
-        vol = convert_stl_to_voxels(all_stl_paths[ind], resolution=100)
-        name = all_stl_paths[ind].removesuffix('.stl').removeprefix('./static/data/solids').removeprefix('/').removeprefix('\\')
-        np.save(f'./static/data/solids/{name}.npy',vol)
+
+        # Converts if it doesn't exist
+        if path.exists(all_stl_paths[ind].removesuffix('.stl') + '.npy'):
+            print(f"{(all_stl_paths[ind].removesuffix('.stl') + '.npy')} has an npy counterpart!")
+
+        else:
+            name = all_stl_paths[ind].removesuffix('.stl').removeprefix('./static/data/solids').removeprefix('/').removeprefix('\\')
+            print(f'Converting model: {name}')
+            vol = convert_stl_to_voxels(all_stl_paths[ind], resolution=100)
+            np.save(f'./static/data/solids/{name}.npy',vol)
